backend/requirements.txt
asgiref==3.7.2
astroid==3.1.0
async-timeout==4.0.3
channels==4.1.0
channels-redis==4.2.0
coverage==7.4.4
dill==0.3.8
daphne==4.1.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-rest-passwordreset==1.4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
isort==5.13.2
mccabe==0.7.0
msgpack==1.0.8
pillow==10.2.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pylint==3.1.0
python-decouple==3.8
pytz==2024.1
redis==5.0.6
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.4
typing_extensions==4.10.0

